requirements.txt
# Core dependencies
fastapi
uvicorn
python-dotenv
pydantic
python-multipart

# Document processing
pyPDF2
python-docx

# Vector stores
faiss-cpu
chromadb

# Database
supabase

# LLM and RAG frameworks
langchain
langchain-community
langchain-core
langchain-openai
openai
llama-index

# Optional: OpenRouter support
requests

# Email functionality
aiosmtplib 
imap-client
supabase
jinja2
email-validator
aiosmtplib

# For development
black
pytest